Renamed `arch_as_root` script to `arch_0`
Now this first installation script doesn't have to be run as root.
Instead, the idea is you'll download this script (e.g.: with `curl`)
and run it straight off. It will then detect whether you're a root user,
and if so create a normal user account as required, then it will run the
rest of the script as that unpriviledged user to improve security, or,
if the current user was already unpriviledged, just exec the rest of the
script.

Then the script will attempt to clone the dotfiles, run `dotf checkout`
automatically, and then run the actual 'install' script
`.config/bin/install/arch`.