Tried to autounlock gnome keyring, moved session launcher into dotfiles
The session launcher is called `fdm`, standing for 'fzf-display
manager'. It is really simple - basically just a bash script which
contains presets for Xorg and Wayland sessions which I like to use.

Also added multi-monitor support to sway config, with Mod+n and Mod+m to
move a workspace between monitors